backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-ninja>=1.3.0",
    "pillow>=11.1.0",
    "ruff>=0.8.5",
]
